Improve root test detection.
Previously, we looked for a “frameElement.id” property to determine if
a document running tests was the “root” test runner or not.

This was too naive as it assumes that integrators looking to iframe a
test suite into a part of a parent application _won’t_ set an “id”
property / attribute on that iframe.

Instead, we now use a “data-*” attribute with a namespace:

```
data-x-test-test-id
```

See #32.