Refactor to split into multiple files.
We previously wanted to have a “single file” for simplicity, but it ends
up being harder to reason about. The new “exports” configuration in
package.json files allows us to indicate which files are actually
internal (versus meant to be imported by integrators).

There’s not such a good case to keep this in a single file anymore.