Add filtering based on “x-test-name” param.
A new `?x-test-name` param is accepted as a query parameter which will
cause `x-test` to go into “filtering” mode. In this mode, any test lines
which do not match the provided name (and empty subtests) will be
omitted from the output stream entirely.

This means that the parameter impacts the browser output and the CLI
output equally. I.e., you can use this filtering in the command line or
via the new input box in the UI.

Closes #58.